refactor(database/online): separate unit tests to two files
